sql_equivalence/requirements.txt
python-dotenv
setuptools
langchain
langchain-community
langchain-ollama
langchain-openai
langchain-anthropic
langchain-google-genai
ollama
pydantic
sqlparse
streamlit
